Port the parser to js (#7)
I want to be able to bootstrap the compiler easily, and it looks like both the JS and Rust parsers are roughly the same speed (based on the parsing of test262/test/**/*.js files).
Still keeping around the old rust parser for now. Will remove it when there's no need for it