[ljs]: Use ResolveResult instead of ResolveImportsResult
In hindsight, resolving imports as a separate phase isn't that helpful
for typechecking. i.e. Having a
ValueBindingExcluding<"Import" | "ImportStarAs"> vs ValueBinding isn't
that helpful. The typechecker has to deal with circularity anyway so
it makes sense to follow the chain of imports to its root declaration
in the typechecker.
This also makes it easier to answer the question "Where should this
diagnostic be reported? In the typechecker or in resolve_imports?".