[ljs]: Replace manual Type.Error calls with emit_error_type
I want to ensure that every Type.Error construction is always
accompanied by a call to `diagnostics.push`.

Secondly, I want to guarantee that the same diagnostic isn't
emitted twice because we accidentally called a function that
emits the diagnostic twice.

To ensure this I'll add a phantom type param to `emit_error_type`
to make sure it's only callable if you have a proof that this
function call is de-duplicated.

This requires further changes to the typechecker.